Feature, Add Quality Option
This patch exposes the `quality` method from ImageMagick.

**Processor**
A new processor was added that can be used to only apply the quality
option without doing any further converting.

**Converter Option**
A new `quality` option was also added to the `Refile::MiniMagick#call`
method. It allows to change the quality while doing other processing at
the same time.

**Specs**
The test suite has been extended to include specs for the new `quality`
converter and for the new and existing options on
`Refile::MiniMagick#call` method.

**Changes to Refile**

Those changes are backwards compatible. However, A second PR to the
refile gem itself would be needed to make the `quality` option usable.

Closes https://github.com/refile/refile-mini_magick/issues/22